third_party/WebKit/LayoutTests/inspector/coverage/segments-merge-expected.txt

Tests the merge of disjoint segment lists in CoverageModel.

A: []
B: []
merged: []
A: [{"end":10,"count":1,"depth":1}]
B: []
merged: [{"end":10,"count":1,"depth":1}]
A: [{"end":10,"count":1,"depth":1}]
B: [{"end":10,"count":1,"depth":1}]
merged: [{"end":10,"count":2,"depth":1}]
A: [{"end":10,"count":1,"depth":1}]
B: [{"end":20,"count":1,"depth":1}]
merged: [{"end":10,"count":2,"depth":1},{"end":20,"count":1,"depth":1}]
A: [{"end":10,"count":1,"depth":1},{"end":20,"count":1,"depth":1}]
B: []
merged: [{"end":10,"count":1,"depth":1},{"end":20,"count":1,"depth":1}]
A: [{"end":30,"count":1,"depth":1}]
B: [{"end":10,"depth":0},{"end":20,"count":2,"depth":1}]
merged: [{"end":10,"count":1,"depth":1},{"end":20,"count":3,"depth":1},{"end":30,"count":1,"depth":1}]
A: [{"end":30,"depth":0}]
B: [{"end":10,"depth":0},{"end":20,"count":2,"depth":1}]
merged: [{"end":10,"depth":0},{"end":20,"count":2,"depth":1},{"end":30,"depth":0}]
